+++ src/main/java/com/accenture/repository/employee/EmployeeRepo.java
@@ -3,5 +3,5 @@ package com.accenture.repository.employee;
 import com.accenture.entity.employee.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepo extends CrudRepository<Employee, Integer> {
+public interface EmployeeRepo extends CrudRepository<Employee, Long> {
 }
