+++ src/main/java/com/accenture/repository/orders/OrdersRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OrdersRepo extends CrudRepository<Orders, Integer> {
+public interface OrdersRepo extends CrudRepository<Orders, Long> {
 
     List<Orders> findByTag(String tag);
 }
